+++ enki/extra/history.md
@@ -1,3 +1,12 @@
 # Enki installation - Extra
 
-To be completes
+This section describes builds that are not described in the BLFS book.
+
+## fastfetch
+
+fastfetch is available in github: [https://github.com/fastfetch-cli/fastfetch](fastfetch). The build instructions are in
+[https://github.com/fastfetch-cli/fastfetch/wiki/Building](fastfetch building).
+
+It requires at least `cmake` (available from BLFS).
+
+We fill also add